fix: fix ReferenceError: require is not defined in ES module scope
An error occurred while running semantic-release:
ReferenceError: require is not defined in ES module scope, you can use import instead
This file is being treated as an ES module because it has a '.js' file extension and
'./node_modules/@amanda-mitchell/semantic-release-npm-multiple/package.json' contains "type": "module".
To treat it as a CommonJS script, rename it to use the '.cjs' file extension.